Separate handler builders from the main logging stuff
